refactor: split file manager ssh helpers
