+++ src/backend/ssh/file-manager-log.ts
@@ -1,0 +1,15 @@
+import type { ConnectionStage, LogEntry } from "../../types/connection-log.js";
+
+export function createConnectionLog(
+  type: "info" | "success" | "warning" | "error",
+  stage: ConnectionStage,
+  message: string,
+  details?: Record<string, unknown>,
+): Omit<LogEntry, "id" | "timestamp"> {
+  return {
+    type,
+    stage,
+    message,
+    details,
+  };
+}
